refactor(Validator & Parser): extract validate function to Validator class from Parser class
